fixed compile warnings on MSVCx64

+++ src/os/linux/interface.hpp
@@ -37,7 +37,7 @@ public:
     bool open ();
     bool close ();
     bool sendPacket (const uint8_t* payload, size_t length, const cTimeval& t);
-    bool prepareSendQueue (int packetCnt, size_t totalBytes, bool synchronized){return true;}
+    bool prepareSendQueue (size_t packetCnt, size_t totalBytes, bool synchronized){return true;}
     bool flushSendQueue (void){return true;}
     bool getMAC (cMacAddress&);
     bool getIPv4 (cIpAddress&);
